+++ CHANGES.md
@@ -4,10 +4,18 @@ Krangl Release History
 v0.4 (Not yet released)
 ----
 
-* spread-gather support for data reshaping (fixes [#2](https://github.com/holgerbrandl/krangl/issues/2))
-* improve reshaping functionality by adding `unite` and `separate` (fixes [#9](https://github.com/holgerbrandl/krangl/issues/9))
+New Features
+
+* `spread()`-`gather()` support for elegant data reshaping (fixes [#2](https://github.com/holgerbrandl/krangl/issues/2))
+* Improve reshaping functionality by adding `unite` and `separate` (fixes [#9](https://github.com/holgerbrandl/krangl/issues/9))
+* Added `sampleFrac()` and `sampleN()` for random sub-sampling of data-frames (either with or without replacement)
+
+Important Bug Fixes
 * `mutate()` can now change existing columns without altering column positions
+
+Other
 * New property accessor  `DataFrame.cols` to access all columns of a data-frame
+
 
 
 v0.3